refactor: move style contants to separate folder
