+++ src/components/pages/parts/SectionBase/styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import spacing from '../../../spacing'
+import spacing from '../../../styles/spacing'
 
 export const SectionContainer = styled.section`
     padding: 0 ${spacing(5)};
